Move DataType into a top-level class
